requirements.txt
requests==2.28.1
python-dotenv==1.0.0
html2text==2024.2.26
pymongo==4.8.0
python-crontab==3.2.0